Fixes related to normal maps and the memory allocator. Improve sampling with cosine weighted importance sampling.

+++ src/post_process/display.glsl
@@ -3,7 +3,7 @@ void main() {
     if (any(greaterThanEqual(pixel_index, constants.screen_size))) {
         return;
     }
-    f16vec3 color = f16vec3(imageLoad(render_target, ivec2(pixel_index)).rgb);
+    vec3 color = vec3(imageLoad(render_target, ivec2(pixel_index)).rgb);
     color = neutral_tonemap(color);
     imageStore(swapchain, ivec2(pixel_index), vec4(vec3(color), 1.0));
 }
